+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Release TBD
   * Replace Orders main menu option, add shipment listing.
   * Allow coupons to be voided by administrators.
   * Create packing lists for each shipment.
+  * Add authorized group to payment gateways, for net terms.
 
 ## v0.7.1
 Release 2019-08-02